### 2025-10-07 #9170 @editorialbot
[PRE REVIEW]: cowfootR: An R Package for Dairy Farm Carbon Footprint
**Submitting author:** @juanmarcosmoreno-arch (<a href="http://orcid.org/0009-0002-8116-2871">Juan Marcos Moreno Moreira</a>)
**Repository:** https://github.com/juanmarcosmoreno-arch/cowfootR
**Branch with paper.md** (empty if default branch): main
**Version:** 0.0.1
