### 2025-10-07 #9170 @juanmarcosmoreno-arch
Hi, thank you for checking.
The cowfootR package has been in development for some time, but this public GitHub repository was created specifically for the JOSS submission. The software was developed and tested internally at Conaprole, Uruguay’s leading dairy cooperative and one of the largest exporters of dairy products in Latin America.

Since this year, cowfootR has been used within Conaprole to estimate the carbon footprint of dairy farms, following IPCC (2019) and IDF (2022) methodologies. The package was recently accepted and published on CRAN, and this repository hosts the public, stable version along with its full documentation. Earlier versions were maintained in internal version control systems before being migrated here to make the tool openly available and citable through JOSS.
