### 2025-12-05 #9170 @jhelvy
Hi @kdorheim, I've never reviewed for JOSS before, but I suppose it's worth giving it a try. Full disclosure: it's the end of the semester, and I've got two other papers in my review queue, so this may need to wait a while, possibly until the new year with holidays coming up. If that's a reasonable time, I could do it.
