### 2025-12-08 #9572 @editorialbot
[REVIEW]: cowfootR: An R Package for Dairy Farm Carbon Footprint Assessment
**Submitting author:** @juanmarcosmoreno-arch (<a href="http://orcid.org/0009-0002-8116-2871">Juan Marcos Moreno Moreira</a>)
**Repository:** https://github.com/juanmarcosmoreno-arch/cowfootR
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.3.0
