### 2026-01-12 #9572 @juanmarcosmoreno-arch
Thank you for the suggestion and the resources. This is my first JOSS submission, and your comments are very helpful for improving my future development workflow. I will include links to relevant commits when responding to issues and adopt a PR-based workflow going forward.
