### 2026-01-22 #9572 @juanmarcosmoreno-arch
thanks @kdorheim , I can suggest to @suarezveirano or @NAlcan who has experience with R packages and veterinary medicine. thanks
