### 2026-02-02 #9572 @kdorheim
@suarezveirano, @NAlcan, & @zarrarkhan would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/pre-2026-submission/. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
