### 2026-02-06 #9572 @zarrarkhan
> [@suarezveirano](https://github.com/suarezveirano), [@NAlcan](https://github.com/NAlcan), & [@zarrarkhan](https://github.com/zarrarkhan) would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/pre-2026-submission/. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.

Hi @kdorheim! This looks very interesting. Thanks for inviting me to review. Unfortunately, I do not have the bandwidth right now to take this on. But will keep an eye on this to test out later! : ).
