### 2026-03-10 #9572 @kendalynnm
This is a nice package that passed all my within software checks. While the utility of the package, it's functionality and usage are well detailed, the development of the literature comparison components of this JOSS submission need some additional development. I'll note that the writing that is present is clear and the references present are appropriately formatted, but left these unchecked as more of both (writing and references) is needed.

Final minor note, the link to this vignette on the repo webpage is broken, as the file does not currently exist in Main:
vignettes/03 - IPCC Methodology Tiers cowfootR.Rmd
