### 2026-03-10 juanmarcosmoreno-arch/cowfootR#7 @kendalynnm
vignette 03 link
On the package webpage: https://juanmarcosmoreno-arch.github.io/cowfootR/index.html

The vignette link at the top of article "IPCC Methodology Tiers in cowfootR" leads to any empty / missing file.
