### 2026-03-10 #9572 @kdorheim
Linking issues related to @kendalynnm's review

* https://github.com/juanmarcosmoreno-arch/cowfootR/issues/7
