### 2026-03-12 999346c
Fix pkgdown article links after vignette rename
