### 2026-03-12 483224e
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@999346c08baf511899a23cdf3b18af03730673a0 🚀
