+++ authors.md
@@ -10,13 +10,13 @@ Source:
 [`DESCRIPTION`](https://github.com/juanmarcosmoreno-arch/cowfootR/blob/main/DESCRIPTION)
 
 Moreno J (2026). *cowfootR: Dairy Farm Carbon Footprint Assessment*. R
-package version 0.1.2,
+package version 0.1.3,
 <https://github.com/juanmarcosmoreno-arch/cowfootR>.
 
     @Manual{,
       title = {cowfootR: Dairy Farm Carbon Footprint Assessment},
       author = {Juan Moreno},
       year = {2026},
-      note = {R package version 0.1.2},
+      note = {R package version 0.1.3},
       url = {https://github.com/juanmarcosmoreno-arch/cowfootR},
     }
