+++ reference/calc_emissions_inputs.md
@@ -177,7 +177,7 @@ calc_emissions_inputs(conc_kg = 1000, fert_n_kg = 200, region = "EU")
 #> [1] "kg CO2e per kg DM"
 #> 
 #> $units$ef_truck
-#> [1] "kg CO2e per (kg·km)"
+#> [1] "kg CO2e per (kg*km)"
 #> 
 #> 
 #> $emissions_breakdown
@@ -314,7 +314,7 @@ calc_emissions_inputs(conc_kg = 1000, fert_n_kg = 200, region = "EU")
 #> [1] 1e-04
 #> 
 #> $emission_factors_used$transport$unit
-#> [1] "kg CO2e per (kg·km)"
+#> [1] "kg CO2e per (kg*km)"
 #> 
 #> $emission_factors_used$transport$transport_km
 #> [1] 0
@@ -391,10 +391,293 @@ calc_emissions_inputs(conc_kg = 1000, fert_n_kg = 200, region = "EU")
 #> [1] "IDF 2022; generic LCI sources"
 #> 
 #> $date
-#> [1] "2026-01-13"
+#> [1] "2026-03-12"
 #> 
 
 # With uncertainty analysis
 calc_emissions_inputs(feed_corn_kg = 2000, region = "US", include_uncertainty = TRUE)
-#> Error in calculate_input_uncertainties(quantities = list(conc_kg = conc_kg,     fert_n_kg = fert_n_kg, plastic_kg = plastic_kg, feeds = list(grain_dry = feed_grain_dry_kg,         grain_wet = feed_grain_wet_kg, ration = feed_ration_kg,         byproducts = feed_byproducts_kg, proteins = feed_proteins_kg,         corn = feed_corn_kg, soy = feed_soy_kg, wheat = feed_wheat_kg)),     factors = list(conc = list(mean = ef_conc, range = ef_conc_range),         fert = list(mean = ef_fert, range = ef_fert_range), plastic = list(mean = ef_plastic,             range = ef_plastic_range), feeds = feed_factors)): could not find function "calculate_input_uncertainties"
+#> $source
+#> [1] "inputs"
+#> 
+#> $units
+#> $units$co2eq_kg
+#> [1] "kg CO2eq yr-1"
+#> 
+#> $units$conc_kg
+#> [1] "kg yr-1"
+#> 
+#> $units$fert_n_kg
+#> [1] "kg N yr-1"
+#> 
+#> $units$plastic_kg
+#> [1] "kg yr-1"
+#> 
+#> $units$feed_kg
+#> [1] "kg DM yr-1"
+#> 
+#> $units$transport_km
+#> [1] "km"
+#> 
+#> $units$ef_conc
+#> [1] "kg CO2e per kg"
+#> 
+#> $units$ef_fert
+#> [1] "kg CO2e per kg N"
+#> 
+#> $units$ef_plastic
+#> [1] "kg CO2e per kg"
+#> 
+#> $units$ef_feed
+#> [1] "kg CO2e per kg DM"
+#> 
+#> $units$ef_truck
+#> [1] "kg CO2e per (kg*km)"
+#> 
+#> 
+#> $emissions_breakdown
+#> $emissions_breakdown$concentrate_co2eq_kg
+#> [1] 0
+#> 
+#> $emissions_breakdown$fertilizer_co2eq_kg
+#> [1] 0
+#> 
+#> $emissions_breakdown$plastic_co2eq_kg
+#> [1] 0
+#> 
+#> $emissions_breakdown$feeds_co2eq_kg
+#>  grain_dry  grain_wet     ration byproducts   proteins       corn        soy 
[…]
