+++ articles/index.md
@@ -3,19 +3,20 @@
 ### Get started
 
 - [Get started with
-  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/00%20-%20Get%20started%20with%20cowfootR.md):
+  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/get-started.md):
 
 ### Core workflows
 
-- [Introduction to Dairy Life Cycle Assessment with
-  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/01%20-%20Introduction%20to%20Dairy%20Life%20Cycle%20Assessment.md):
+- [Introduction to Dairy Life Cycle
+  Assessment](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/introduction-lca.md):
 - [Single Farm Analysis
-  (cowfootR)](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/02%20-%20Single%20Farm%20Analysis%20(cowfootR).md):
+  (cowfootR)](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/single-farm-analysis.md):
 - [IPCC Methodology Tiers in
-  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/03%20-%20IPCC%20Methodology%20Tiers%20cowfootR.md):
+  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/ipcc-tiers.md):
 
 ### Reference
 
-- [Complete_Parameter_Reference_Guide](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/04%20-%20Complete%20Parameter%20Reference%20Guide.md):
+- [Complete Parameter Reference
+  Guide](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/parameter-reference.md):
 - [cowfootR
   Workflow](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/workflow.md):
