### 2026-03-12 cbbe21b
Revise environmental impact section and introduce cowfootR
Revised the section on the environmental impact of milk production, emphasizing the carbon footprint and its significance in greenhouse gas emissions. Updated details on existing methodologies and introduced the cowfootR package as a solution for standardized carbon assessment.
