### 2026-03-12 6991f9a
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@cbbe21b38e494a19f3ffb4566c7e16dd9af82232 🚀
