+++ reference/export_hdc_report.md
@@ -74,7 +74,7 @@ class(br) <- "cf_batch_complete"
 
 f <- tempfile(fileext = ".xlsx")
 export_hdc_report(br, file = f)
-#> Batch report saved to: /tmp/RtmpFw5luG/file1ade55bf186f.xlsx
+#> Batch report saved to: /tmp/Rtmpu7pSLK/file1a137ad1435e.xlsx
 file.exists(f)
 #> [1] TRUE
 # }
