### 2026-03-12 4bbf904
Add citation for Poore and Nemecek article
