+++ paper.bib
@@ -54,3 +54,13 @@
   address = {Strasbourg, France},
   note = {Text with EEA Relevance}
 }
+@article{poore2018,
+  title = {Reducing food’s environmental impacts through producers and consumers},
+  author = {Poore, J. and Nemecek, T.},
+  journal = {Science},
+  volume = {360},
+  number = {6392},
+  pages = {987--992},
+  year = {2018},
+  doi = {https://dx.doi.org/10.1126/science.aaq0216}
+}
