### 2026-03-12 97e1b8c
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@4bbf9048b6e4327c1a303db7f404280c82727e6f 🚀
