### 2026-03-12 c3f799e
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@2e8f1220332b4a564dc0503735d57b842f521869 🚀
