### 2026-03-20 d7a7caf
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@2e8f1220332b4a564dc0503735d57b842f521869 🚀
