+++ articles/single-farm-analysis.md
@@ -502,7 +502,7 @@ manure_emissions
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2026-03-12"
+#> [1] "2026-03-20"
 #> 
 #> $per_cow
 #> $per_cow$ch4_kg
@@ -633,7 +633,7 @@ soil_emissions
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2026-03-12"
+#> [1] "2026-03-20"
 #> 
 #> $per_hectare_metrics
 #> $per_hectare_metrics$n_input_kg_per_ha
@@ -788,7 +788,7 @@ energy_emissions
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2026-03-12"
+#> [1] "2026-03-20"
 #> 
 #> $energy_metrics
 #> $energy_metrics$electricity_share_pct
@@ -1070,7 +1070,7 @@ input_emissions
 #> [1] "IDF 2022; generic LCI sources"
 #> 
 #> $date
-#> [1] "2026-03-12"
+#> [1] "2026-03-20"
 ```
 
 ### Step 4: Total emissions and hotspot analysis
@@ -1103,7 +1103,7 @@ total_emissions
 #>   manure : 347959.2 kg CO2eq yr-1 
 #>   soil : 112601.8 kg CO2eq yr-1 
 #> 
-#> Calculated on: 2026-03-12
+#> Calculated on: 2026-03-20
 ```
 
 #### Visualize emission sources (bar chart)
@@ -1166,7 +1166,7 @@ milk_intensity
 #>  Protein content: 3.3 %
 #> 
 #> Total emissions: 1,590,294 kg CO2eq
-#> Calculated on: 2026-03-12
+#> Calculated on: 2026-03-20
 ```
 
 #### Area Intensity (kg CO₂eq per ha)
@@ -1209,7 +1209,7 @@ area_intensity
 #>  woodland: 7.0 ha (3.5%) -> 55660 kg CO2eq
 #> 
 #> Total emissions: 1,590,294 kg CO2eq
-#> Calculated on: 2026-03-12
+#> Calculated on: 2026-03-20
 ```
 
 #### Benchmarking (interpretation note)
