### 2026-03-20 cce3779
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@4fee494af43709cdac9c448588681180af4a7e4b 🚀
