### 2026-03-20 e2c24a4
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@d50fa635b778f5632f5e8dcac830cd33c85664a8 🚀
