+++ pkgdown.yml
@@ -8,7 +8,7 @@ articles:
   parameter-reference: parameter-reference.html
   single-farm-analysis: single-farm-analysis.html
   workflow: workflow.html
-last_built: 2026-03-20T16:00Z
+last_built: 2026-03-20T16:02Z
 urls:
   reference: https://juanmarcosmoreno-arch.github.io/cowfootR/reference
   article: https://juanmarcosmoreno-arch.github.io/cowfootR/articles
