### 2026-03-20 #9572 @juanmarcosmoreno-arch
All post-review author tasks have been completed:

- Authors, affiliations, and ORCIDs have been carefully checked and are correct.
- A final release of the software has been created: v0.1.3
- https://github.com/juanmarcosmoreno-arch/cowfootR/releases/tag/v0.1.3
- The software has been archived on Zenodo and assigned a DOI:
  https://doi.org/10.5281/zenodo.19135408
- The title and author list in the Zenodo archive match those in the JOSS paper.
- The license in the archive is consistent with the repository license.

Please let me know if anything else is needed.
