### 2026-03-21 #9572 @kdorheim
@juanmarcosmoreno-arch thanks for your promptness I am sorry I was premature in asking for you to do the release. Could you please address the invalid dois in the paper.bib?
