### 2026-03-21 552dd5c
Remove 'https://' from DOIs in paper.bib
