+++ paper.bib
@@ -5,7 +5,7 @@
   volume = {16},
   pages = {420--430},
   year = {2011},
-  doi = {https://doi.org/10.1007/s11367-011-0283-9}
+  doi = {10.1007/s11367-011-0283-9}
 }
 
 @techreport{idf2022,
@@ -34,7 +34,7 @@
   volume = {11},
   pages = {109--118},
   year = {2012},
-  doi = {https://doi.org/10.4081/ijas.2012.e20}
+  doi = {10.4081/ijas.2012.e20}
 }
 
 @article{stolarski2025,
@@ -44,7 +44,7 @@
   volume = {15},
   pages = {8446},
   year = {2025},
-  doi = {https://doi.org/10.3390/app15158446}
+  doi = {10.3390/app15158446}
 }
 
 @misc{eu2022,
@@ -62,5 +62,5 @@
   number = {6392},
   pages = {987--992},
   year = {2018},
-  doi = {https://dx.doi.org/10.1126/science.aaq0216}
+  doi = {10.1126/science.aaq0216}
 }
