### 2026-03-21 #9572 @juanmarcosmoreno-arch
Thanks for the clarification. The DOIs in `paper.bib` have been corrected by removing the URL prefixes and now follow the expected format.
