### 2026-03-21 e6d3b7c
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@552dd5c4d5487fa55ffe337d9301039a119ca435 🚀
