+++ articles/introduction-lca.md
@@ -307,7 +307,7 @@ enteric_emissions
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2026-03-20"
+#> [1] "2026-03-21"
 #> 
 #> $per_animal
 #> $per_animal$ch4_kg
@@ -428,7 +428,7 @@ manure_emissions
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2026-03-20"
+#> [1] "2026-03-21"
 #> 
 #> $per_cow
 #> $per_cow$ch4_kg
@@ -567,7 +567,7 @@ soil_emissions
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2026-03-20"
+#> [1] "2026-03-21"
 #> 
 #> $per_hectare_metrics
 #> $per_hectare_metrics$n_input_kg_per_ha
@@ -729,7 +729,7 @@ energy_emissions
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2026-03-20"
+#> [1] "2026-03-21"
 #> 
 #> $energy_metrics
 #> $energy_metrics$electricity_share_pct
@@ -1012,7 +1012,7 @@ input_emissions
 #> [1] "IDF 2022; generic LCI sources"
 #> 
 #> $date
-#> [1] "2026-03-20"
+#> [1] "2026-03-21"
 ```
 
 #### Step 4: Aggregate Total Emissions
@@ -1046,7 +1046,7 @@ total_emissions
 #>   manure : 159755.4 kg CO2eq yr-1 
 #>   soil : 69658.88 kg CO2eq yr-1 
 #> 
-#> Calculated on: 2026-03-20
+#> Calculated on: 2026-03-21
 ```
 
 #### Step 5: Calculate Intensity Metrics
@@ -1080,7 +1080,7 @@ milk_intensity
 #>  Protein content: 3.2 %
 #> 
 #> Total emissions: 645,948 kg CO2eq
-#> Calculated on: 2026-03-20
+#> Calculated on: 2026-03-21
 ```
 
 ##### Area Intensity
@@ -1117,7 +1117,7 @@ area_intensity
 #>  infrastructure: 5.0 ha (4.2%) -> 26914 kg CO2eq
 #> 
 #> Total emissions: 645,948 kg CO2eq
-#> Calculated on: 2026-03-20
+#> Calculated on: 2026-03-21
 ```
 
 ### Visualizing Results
