### 2026-03-21 #9572 @kdorheim
@juanmarcosmoreno-arch things are looking good! Could you please do another release? Share the new version number and Zenodo DOI?
