### 2026-03-21 bfb0722
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@1591223bdf91154eff1b6c814e39fac6e3eee646 🚀
