+++ reference/export_hdc_report.md
@@ -74,7 +74,7 @@ class(br) <- "cf_batch_complete"
 
 f <- tempfile(fileext = ".xlsx")
 export_hdc_report(br, file = f)
-#> Batch report saved to: /tmp/RtmpdnAskB/file1c52124af36e.xlsx
+#> Batch report saved to: /tmp/Rtmp9BLesg/file1c78301d685f.xlsx
 file.exists(f)
 #> [1] TRUE
 # }
