### 2026-03-21 #9572 @juanmarcosmoreno-arch
Thanks for the feedback.

A new release has been created with the requested fixes:

Zenodo DOI: https://doi.org/10.5281/zenodo.19152615

Please let me know if anything else is needed.
