### 2026-03-27 #9572 @juanmarcosmoreno-arch
Hi @kdorheim, thanks. Has the paper been published yet?
