### 2026-03-30 #9572 @kdorheim
@juanmarcosmoreno-arch thanks for your patience, we are just waiting to hear from @kthyng, the editor in charge.

@kthyng I think the only outstanding problem issue at the moment is the related to the DOI link
