### 2026-04-17 #9572 @juanmarcosmoreno-arch
@kthyng Thanks for the suggestion. I agree with aligning the metadata, but unfortunately I’m not able to modify the Zenodo.

Please let me know if you would prefer that I create a new version to update the metadata, although this would generate a new DOI. Otherwise, I would prefer to keep it as is.
