### 2026-05-07 #9572 @juanmarcosmoreno-arch
Thank you very much to the editor and reviewers for the time, feedback, and suggestions throughout the review process.

The comments and recommendations greatly improved both the package and the manuscript. I truly appreciate the constructive and supportive review experience.

I’m very happy to see cowfootR accepted in JOSS.
